Improve test validator and add ability to define port
