+++ crates/sui-test-data/Cargo.toml
@@ -1,39 +1,23 @@
 [package]
-name = "sui-open-rpc"
-version = "0.6.1"
+name = "sui-test-data"
+version = "0.0.1"
 authors = ["Mysten Labs <build@mystenlabs.com>"]
 license = "Apache-2.0"
 publish = false
 edition = "2021"
 
 [dependencies]
-schemars = "0.8.10"
-serde = "1.0.141"
-serde_json = "1.0.83"
-workspace-hack = { path = "../workspace-hack"}
-
-[dev-dependencies]
+serde = "1.0.140"
 anyhow = { version = "1.0.58", features = ["backtrace"] }
-clap = { version = "3.1.17", features = ["derive"] }
-pretty_assertions = "1.2.0"
-serde_json = "1.0.83"
-tokio = { version = "1.20.1", features = ["full"] }
+serde_json = "1.0.80"
 hyper = { version = "0.14.20", features = ["full"] }
-
 sui = { path = "../sui" }
 sui-json-rpc = { path = "../sui-json-rpc" }
 sui-json-rpc-types = { path = "../sui-json-rpc-types" }
 sui-json = { path = "../sui-json" }
 sui-types = { path = "../sui-types" }
 sui-config = { path = "../sui-config" }
-sui-test-data = { path = "../sui-test-data" }
 test-utils = { path = "../test-utils" }
-rand = "0.7.3"
+workspace-hack = { path = "../workspace-hack"}
 
 move-package = { git = "https://github.com/move-language/move", rev = "79071528524f08b12e9abb84c1094d8e976aa17a" }
-move-core-types = { git = "https://github.com/move-language/move", rev = "79071528524f08b12e9abb84c1094d8e976aa17a", features = ["address20"] }
-
-[[example]]
-name = "generate-json-rpc-spec"
-path = "src/generate_json_rpc_spec.rs"
-test = false
