+++ crates/sui-test-validator/Cargo.toml
@@ -25,6 +25,7 @@ sui-json-rpc = { path = "../sui-json-rpc" }
 sui-json-rpc-types = { path = "../sui-json-rpc-types" }
 sui-json = { path = "../sui-json" }
 sui-types = { path = "../sui-types" }
+sui-test-data = { path = "../sui-test-data" }
 sui-config = { path = "../sui-config" }
 test-utils = { path = "../test-utils" }
