+++ crates/test-utils/src/network.rs
@@ -114,8 +114,10 @@ pub async fn setup_network_and_wallet() -> Result<(Swarm, WalletContext, SuiAddr
 
 async fn start_rpc_gateway(
     config_path: &Path,
+    port: Option<u16>,
 ) -> Result<(SocketAddr, HttpServerHandle), anyhow::Error> {
-    let server = HttpServerBuilder::default().build("127.0.0.1:0").await?;
+    let server_url = format!("127.0.0.1:{}", port.unwrap_or(0));
+    let server = HttpServerBuilder::default().build(server_url).await?;
     let addr = server.local_addr()?;
 
     let config = PersistedConfig::read(config_path)?;
@@ -133,18 +135,20 @@ async fn start_rpc_gateway(
 pub async fn start_rpc_test_network(
     genesis_config: Option<GenesisConfig>,
 ) -> Result<TestNetwork, anyhow::Error> {
-    start_rpc_test_network_with_fullnode(genesis_config, 0, None).await
+    start_rpc_test_network_with_fullnode(genesis_config, 0, None, None).await
 }
 
 pub async fn start_rpc_test_network_with_fullnode(
     genesis_config: Option<GenesisConfig>,
     fullnode_count: usize,
     config_dir: Option<&Path>,
+    port: Option<u16>,
 ) -> Result<TestNetwork, anyhow::Error> {
-    let network = start_test_network_with_fullnodes(genesis_config, fullnode_count, config_dir).await?;
+    let network =
+        start_test_network_with_fullnodes(genesis_config, fullnode_count, config_dir).await?;
     let working_dir = network.dir();
     let (server_addr, rpc_server_handle) =
-        start_rpc_gateway(&working_dir.join(SUI_GATEWAY_CONFIG)).await?;
+        start_rpc_gateway(&working_dir.join(SUI_GATEWAY_CONFIG), port).await?;
     let mut wallet_conf: SuiClientConfig =
         PersistedConfig::read(&working_dir.join(SUI_CLIENT_CONFIG))?;
     let rpc_url = format!("http://{}", server_addr);
